Add data structure for link_data
The link_data object is responsible for the 1..* relationship
between terms and related links. The commit introduces the object
and a suite of tests that dynamically test the link_data object for
referential integrity and the types of properties. This approach
has been taken to ensure that any errors are picked up when new links
are added.